# spec/openapi.yaml
openapi: 3.1.0
info:
  title: Гармония (Harmony) API
  version: 1.0.0
  summary: REST and realtime API for the Harmony social network.
  description: |
    # Гармония API

    A social network API. All responses share one envelope so a client has a
    single code path for success and failure:

    ```json
    { "ok": true,  "data": { }, "meta": { "count": 15, "has_more": true, "next_cursor": "eyJ..." } }
    { "ok": false, "error": { "code": "not_found", "message": "Пост не найден.", "request_id": "..." } }
    ```

    ## Authentication

    After `POST /auth/login` the API issues two cookies:

    | Cookie | HttpOnly | Purpose |
    |---|---|---|
    | `harmony_access` | yes | Access token (JWT) |
    | `harmony_refresh` | yes | Refresh token, scoped to `/api/v1/auth` |
    | `harmony_csrf` | **no** | CSRF token for the double-submit pattern |

    The access cookie is `HttpOnly` so that an XSS payload cannot read it.
    The CSRF cookie is deliberately readable, because the client must echo it.

    **Every unsafe method** (`POST`, `PATCH`, `PUT`, `DELETE`) must send the CSRF
    token in the `X-CSRF-Token` header, matching the cookie. A correct `Origin`
    or `Referer` is also accepted, which keeps non-browser clients working.

    Non-browser clients may instead send `Authorization: Bearer <access_token>`.

    ## Resources

    Resources are addressed by an opaque 32-character `public_id`, never by the
...
    bugs easier to exploit.

    ## Pagination

...

...

...

...

...

...

...

  contact:
...
servers:
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /legal/documents/{slug}:
    get:
      tags: [Privacy]
      summary: One legal document
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string, enum: [terms, privacy] }
      responses:
        "200": { description: Document }
        "404": { description: Not found }
...
  /me/data/export:
    post:
      tags: [Privacy]
      summary: Request a downloadable archive
      responses:
        "200": { description: Archive ready }
        "409": { description: A request of this kind is already open }
  /me/data/export/{request_id}:
    get:
      tags: [Privacy]
      summary: Download a generated archive
      parameters:
        - name: requestId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: JSON file }
        "404": { description: Not found }
        "409": { description: Not ready, or the download window has closed }
...
  /me/deletion:
    get:
      tags: [Privacy]
      summary: Deletion status
      responses:
        "200": { description: Status }
    post:
      tags: [Privacy]
      summary: Schedule account deletion (Art. 17)
      description: |
        Requires a `reauth_token` from `/auth/reauthenticate`, so a hijacked
        session alone cannot destroy an account. The account becomes unusable
        immediately; the data is erased when the grace period elapses unless
        the deletion is cancelled.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reauth_token]
              properties:
                reauth_token: { type: string }
                reason: { type: string }
      responses:
        "200": { description: Scheduled }
        "403": { description: Step-up token missing or invalid }
        "409": { description: Already scheduled }
  /me/deletion/cancel:
    post:
      tags: [Privacy]
      summary: Cancel a scheduled deletion
      responses:
        "200": { description: Cancelled }
        "409": { description: Nothing to cancel }

  # --------------------------------------------------------------- Admin ---
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: harmony_access
      description: |
        Access token set by `/auth/login`. Send `X-CSRF-Token` on unsafe methods.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: For non-browser clients. Exempt from CSRF.
...